Refactor RPC and Game types to remove PerformActionCommand and define PerformActionResponse structure
