+++ sdk/src/types/game.ts
@@ -169,3 +169,32 @@ export type GameOptionsResponse = {
     address: string;
     gameOptions: GameOptionsDTO;
 };
+
+
+export type TransactionResponse = {
+    nonce: string;
+    to: string;
+    from: string;
+    value: string;
+    hash: string;
+    signature: string;
+    timestamp: string;
+    data?: string;
+}
+
+export type GameStateResponse = {
+    state: TexasHoldemStateDTO;
+}
+
+// export type PerformActionResponse = GameStateResponse & TransactionResponse;
+export type PerformActionResponse= {
+    state: TexasHoldemStateDTO;
+    nonce: string;
+    to: string;
+    from: string;
+    value: string;
+    hash: string;
+    signature: string;
+    timestamp: string;
+    data?: string;
+}
